+++ src/livery.h
@@ -78,8 +78,8 @@ enum LiveryClass {
 /** Information about a particular livery. */
 struct Livery {
 	byte in_use;  ///< Bit 0 set if this livery should override the default livery first colour, Bit 1 for the second colour.
-	byte colour1; ///< First colour, for all vehicles.
-	byte colour2; ///< Second colour, for vehicles with 2CC support.
+	uint32 colour1; ///< First colour, for all vehicles.
+	uint32 colour2; ///< Second colour, for vehicles with 2CC support.
 
 	PaletteID cached_pal_1cc; ///< NOSAVE: cached 1CC palette.
 	PaletteID cached_pal_2cc; ///< NOSAVE: cached 2CC palette.
